Source files added!
CMakeLists.txt: A type of Makefile to be used with the CMake program to
build project files.
Isolines.cxx: The source code for the project.
Isolines.vtk: The data that the final executable needs in order to run.
This file MUST be moved to the same location as the executable.